Temporarily factorize parse_run initialization code into UsesParseRun
